+++ core/ratarmountcore/compressions.py
@@ -6,7 +6,9 @@ import concurrent.futures
 import os
 import struct
 import sys
-from typing import IO, Optional, Tuple
+from typing import Callable, IO, Iterable, List, Optional, Tuple
+
+from .utils import isLatinAlpha, isLatinDigit, isLatinHexAlpha, formatNumber, ALPHA, DIGITS, HEX
 
 try:
     import indexed_bzip2
@@ -49,7 +51,7 @@ except ImportError:
 
 # The file object returned by ZipFile.open is not seekable in Python 3.6 for some reason.
 # Therefore disable ZIP support there!
-# I don't see it documented, instead I tested different Python versions with Docker.
+# I don't see it documented, instead, I tested different Python versions with Docker.
 if sys.version_info[0] == 3 and sys.version_info[1] > 6:
     import zipfile
 else:
@@ -134,6 +136,89 @@ def stripSuffixFromTarFile(path: str) -> str:
         path = path[:-4]
 
     return path
+
+
+def hasMatchingAlphabets(a: str, b: str):
+    return (
+        (isLatinAlpha(a) and isLatinAlpha(b))
+        or (isLatinDigit(a) and isLatinDigit(b))
+        or (isLatinHexAlpha(a) and isLatinHexAlpha(b))
+    )
+
+
+def checkForSequence(extensions: Iterable[str], numberFormatter: Callable[[int], str]) -> List[str]:
+    suffixSequence: List[str] = []
+    i = 0
+    suffixLength = len(numberFormatter(0))
+
+    while True:
+        suffix = numberFormatter(i)
+        if suffix in extensions:
+            suffixSequence.append(suffix)
+        elif i > 0 or len(suffix) != suffixLength:
+            # We allow the extensions to start with 0 or 1.
+            # So, even if the zeroth does not exist, do not break, instead also test 1.
+            break
+        i += 1
+
+    return suffixSequence
+
+
+def checkForSplitFile(path: str) -> Optional[Tuple[List[str], str]]:
+    """
+    Returns the paths to all files belonging to the split and a string identifying the format.
+    The latter is one of: '', 'x', 'a' to specify the numbering system: decimal, hexadecimal, alphabetical.
+    The width and starting number can be determined from the extension of the first file path,
+    which is returned in the first member of the tuple.
+    """
+
+    # Check for split files. Note that GNU coreutils' split 8.32 by default creates files in the form:
+    #   xaa, xab, xac, ..., xba, ...
+    # However, for now, don't support those. At least a dot should be used to reduce misidentification!
+    # Split supports not only alphabetical, but also decimal, and hexadecimal numbering schemes and the
+    # latter, both, start at 0, which is also quite unusual.
+    # Most files I encountered in the wild used decimal suffixes starting with name.001.
+    # Split can be customized a lot with these options:
+    #   split [OPTION]... [FILE [PREFIX]]
+    #   --suffix-length
+    #   --additional-suffix=SUFFIX
+    #   --numeric-suffixes[=FROM]
+    #   --hex-suffixes[=FROM]
+    #   --suffix-length
+    # -> It seems like there is no way to specify alphabetical counting to begin at something other than a, aa, ...
+    # To avoid false positives, these restrictions to split's options are made:
+    #   - FROM is 0 or 1. Anything else is too aberrant to be supported.
+    #   - PREFIX ends with a dot (.).
+
+    # These character tests are necessary because Python's built-in isalpha, isdigit and such
+    # all also return true for a lot of Unicode alternatives like for the Thai zero.
+
+    folder, filename = os.path.split(path)  # Get file extensions
+    if '.' not in filename:
+        return None
+    basename, extension = filename.rsplit('.', maxsplit=1)
+
+    # Collect all other files in the folder that might belong to the same split.
+    extensions = [name[len(basename) + 1 :] for name in os.listdir(folder) if name.startswith(basename + '.')]
+    extensions = [e for e in extensions if hasMatchingAlphabets(e, extension)]
+    if not extensions:
+        return None
+    assert extension in extensions
+
+    # Note that even if something consists only of letters or only digits it still might be hexadecimal encoding!
+    maxFormatSpecifier = ''
+    maxExtensions: List[str] = []
+    for formatSpecifier, baseDigits in [('a', ALPHA), ('0', DIGITS), ('x', HEX)]:
+        extensionSequence = checkForSequence(extensions, lambda i: formatNumber(i, baseDigits, len(extension)))
+        if len(extensionSequence) > len(maxExtensions):
+            maxFormatSpecifier = formatSpecifier
+            maxExtensions = extensionSequence
+
+    if maxFormatSpecifier and len(maxExtensions) > 1:
+        paths = [path.rsplit('.', maxsplit=1)[0] + '.' + extension for extension in maxExtensions]
+        return paths, maxFormatSpecifier
+
+    return None
 
 
 def _compressZstd(data):