+++ core/ratarmountcore/factory.py
@@ -7,24 +7,40 @@ import traceback
 
 from typing import IO, Union
 
-from .compressions import supportedCompressions, rarfile, zipfile
+from .compressions import supportedCompressions, checkForSplitFile, rarfile, zipfile
 from .utils import CompressionError, RatarmountError
 from .MountSource import MountSource
 from .FolderMountSource import FolderMountSource
 from .RarMountSource import RarMountSource
+from .SingleFileMountSource import SingleFileMountSource
 from .SQLiteIndexedTar import SQLiteIndexedTar
+from .StenciledFile import JoinedFile
 from .ZipMountSource import ZipMountSource
 
 
 def openMountSource(fileOrPath: Union[str, IO[bytes]], **options) -> MountSource:
     printDebug = int(options.get("printDebug", 0)) if isinstance(options.get("printDebug", 0), int) else 0
 
+    joinedFileName = ''
     if isinstance(fileOrPath, str):
         if not os.path.exists(fileOrPath):
             raise RatarmountError(f"Mount source does not exist: {fileOrPath}")
 
         if os.path.isdir(fileOrPath):
             return FolderMountSource('.' if fileOrPath == '.' else os.path.realpath(fileOrPath))
+
+        splitFileResult = checkForSplitFile(fileOrPath)
+        if splitFileResult:
+            # TODO what about closing files :/?
+            # TODO what if there are a lot of split files, more than the allowed number of opened handles!?
+            # I guess both of these problems can be solved by creating a specialized JoinedFile class :/
+            # That class would have to close all files not currently in the seek/read area and the input would
+            # be a path list not an object list.
+            filesToJoin = splitFileResult[0]
+            joinedFileName = os.path.basename(filesToJoin[0]).rsplit('.', maxsplit=1)[0]
+            if 'indexFilePath' not in options or not options['indexFilePath']:
+                options['indexFilePath'] = filesToJoin[0] + ".index.sqlite"
+            fileOrPath = JoinedFile([open(file, 'rb') for file in filesToJoin])
 
     try:
         if 'rarfile' in sys.modules and rarfile.is_rarfile(fileOrPath):
@@ -84,5 +100,8 @@ def openMountSource(fileOrPath: Union[str, IO[bytes]], **options) -> MountSource
         finally:
             if hasattr(fileOrPath, 'seek'):
                 fileOrPath.seek(0)  # type: ignore
+
+    if joinedFileName and not isinstance(fileOrPath, str):
+        return SingleFileMountSource(joinedFileName, fileOrPath)
 
     raise CompressionError(f"Archive to open ({str(fileOrPath)}) has unrecognized format!")