+++ core/tests/test_utils.py
@@ -9,7 +9,18 @@ import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from ratarmountcore.utils import ceilDiv, LRUCache, Prefetcher  # noqa: E402
+from ratarmountcore.utils import (  # noqa: E402
+    ceilDiv,
+    formatNumber,
+    isLatinAlpha,
+    isLatinDigit,
+    isLatinHexAlpha,
+    ALPHA,
+    DIGITS,
+    HEX,
+    LRUCache,
+    Prefetcher,
+)
 
 
 def test_ceilDiv():
@@ -114,3 +125,102 @@ def test_LRUCache():
     assert len(cache) == 0
     assert 7 not in cache
     assert not cache.keys()
+
+
+def test_isLatinAlpha():
+    assert ALPHA == 'abcdefghijklmnopqrstuvwxyz'
+    assert isLatinAlpha(ALPHA)
+
+    assert isLatinAlpha("a")
+    assert isLatinAlpha("ab")
+    assert isLatinAlpha("z")
+    assert not isLatinAlpha("0")
+    assert not isLatinAlpha("")
+
+    # Python's isalpha is too Unicode. I only want to match the usual
+    # ASCII Latin alphabet not mathematical or Greek symbols!
+    # https://util.unicode.org/UnicodeJsps/confusables.jsp
+    # https://www.unicode.org/Public/security/revision-03/confusablesSummary.txt
+    assert "α".isalpha()
+    assert not isLatinAlpha("α")
+
+    assert "𝒻".isalpha()
+    assert not isLatinAlpha("𝒻")
+
+
+def test_isLatinDigit():
+    assert DIGITS == '0123456789'
+    assert isLatinDigit(DIGITS)
+
+    assert isLatinDigit("0")
+    assert isLatinDigit("10")
+    assert isLatinDigit("9")
+    assert not isLatinDigit("a")
+    assert not isLatinDigit("")
+
+    # Python's isdigit is too Unicode. I only want to match the usual
+    # ASCII Latin digits not decimal digits from other languages!
+    # https://www.compart.com/en/unicode/category/Nd
+    assert "௫".isdigit()
+    assert not isLatinDigit("௫")
+
+    assert "߂".isdigit()
+    assert not isLatinDigit("߂")
+
+
+def test_isLatinHexAlpha():
+    assert HEX == '0123456789abcdef'
+    assert isLatinHexAlpha(HEX)
+
+    assert isLatinHexAlpha("0")
+    assert isLatinHexAlpha("10")
+    assert isLatinHexAlpha("9")
+    assert isLatinHexAlpha("a")
+    assert isLatinHexAlpha("f")
+    assert not isLatinHexAlpha("g")
+    assert not isLatinHexAlpha("")
+
+    # Python's isdigit is too Unicode. I only want to match the usual
+    # ASCII Latin digits not decimal digits from other languages!
+    # https://www.compart.com/en/unicode/category/Nd
+    assert "௫".isdigit()
+    assert not isLatinHexAlpha("௫")
+
+    assert "߂".isdigit()
+    assert not isLatinHexAlpha("߂")
+
+
+def test_formatNumber():
+    assert formatNumber(0, ALPHA) == 'a'
+    assert formatNumber(1, ALPHA) == 'b'
+    assert formatNumber(25, ALPHA) == 'z'
+    assert formatNumber(26, ALPHA) == 'ba'
+    assert formatNumber(26 * 26, ALPHA) == 'baa'
+    assert formatNumber(26 * 26 + 3, ALPHA) == 'bad'
+
+    assert formatNumber(0, ALPHA, 3) == 'aaa'
+    assert formatNumber(1, ALPHA, 3) == 'aab'
+    assert formatNumber(25, ALPHA, 3) == 'aaz'
+    assert formatNumber(26, ALPHA, 3) == 'aba'
+    assert formatNumber(26 * 26, ALPHA, 3) == 'baa'
+    assert formatNumber(26 * 26 + 3, ALPHA, 3) == 'bad'
+
+    assert formatNumber(0, HEX) == '0'
+    assert formatNumber(1, HEX) == '1'
+    assert formatNumber(9, HEX) == '9'
+    assert formatNumber(10, HEX) == 'a'
+    assert formatNumber(15, HEX) == 'f'
+    assert formatNumber(16, HEX) == '10'
+    assert formatNumber(16 * 16, HEX) == '100'
+    assert formatNumber(16 * 16 + 3, HEX) == '103'
+
+    assert formatNumber(0, HEX, 3) == '000'
+    assert formatNumber(1, HEX, 3) == '001'
+    assert formatNumber(9, HEX, 3) == '009'
+    assert formatNumber(10, HEX, 3) == '00a'
+    assert formatNumber(15, HEX, 3) == '00f'
+    assert formatNumber(16, HEX, 3) == '010'
+    assert formatNumber(16 * 16, HEX, 3) == '100'
+    assert formatNumber(16 * 16 + 3, HEX, 3) == '103'
+
+    assert formatNumber(357641610, DIGITS) == '357641610'