+++ tests/create-tests.sh
@@ -258,3 +258,10 @@ for (( i=1; i < 100; ++i )); do
     'rm' "ufo_$( printf %02i "$(( i-1 ))" ).gz"
 done
 mv ufo_99.gz compressed-100-times.gz
+
+
+# Split files
+echo foo >> simple-file-split.001
+echo bar >> simple-file-split.002
+
+split --numeric-suffixes=1 --number=2 --suffix-length=3 single-file.tar single-file-split.tar.